requirements.txt
pytest==7.4.3
pytest-watch==4.2.0
psycopg2-binary==2.9.9
bcrypt==4.1.1
pyyaml==6.0.1